requirements-ds.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile requirements-ds.in
#
cachetools==5.0.0
    # via google-auth
catboost==0.26.1
    # via -r requirements-ds.in
certifi==2021.10.8
    # via requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via pip-tools
cloudpickle==2.0.0
    # via shap
cycler==0.11.0
    # via matplotlib
distlib==0.3.4
    # via virtualenv
filelock==3.6.0
    # via virtualenv
google-api-core[grpc]==2.7.1
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.6.2
    # via
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
    #   pandas-gbq
    #   pydata-google-auth
google-auth-oauthlib==0.5.1
    # via
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-aiplatform==1.9.0
    # via -r requirements-ds.in
google-cloud-bigquery[bqstorage,pandas]==2.26.0
    # via
    #   -r requirements-ds.in
    #   google-cloud-aiplatform
    #   pandas-gbq
google-cloud-bigquery-storage==2.13.0
    # via google-cloud-bigquery
google-cloud-core==2.2.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-storage==1.42.2
    # via
    #   -r requirements-ds.in
    #   google-cloud-aiplatform
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.56.0
    # via
    #   google-api-core
    #   grpcio-status
graphviz==0.19.1
    # via catboost
grpcio==1.45.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   grpcio-status
grpcio-status==1.45.0
    # via google-api-core
identify==2.4.12
    # via pre-commit
idna==3.3
    # via requests
importlib-metadata==4.12.0
    # via
    #   click
    #   pep517
    #   pre-commit
    #   virtualenv
joblib==1.1.0
    # via scikit-learn
kiwisolver==1.4.2
    # via matplotlib
lightgbm==3.3.2
    # via -r requirements-ds.in
llvmlite==0.38.0
    # via numba
loguru==0.5.3
    # via -r requirements-ds.in
matplotlib==3.4.3
    # via
    #   -r requirements-ds.in
    #   catboost
    #   seaborn
nodeenv==1.6.0
    # via pre-commit
numba==0.55.1
    # via shap
numpy==1.21.5
    # via
    #   catboost
    #   lightgbm
    #   matplotlib
    #   numba
    #   pandas
    #   pandas-gbq
    #   patsy
    #   pot
    #   pyarrow
    #   scikit-learn
    #   scipy
    #   seaborn
    #   shap
    #   statsmodels
    #   xgboost
oauthlib==3.2.0
    # via requests-oauthlib
packaging==21.3
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   statsmodels
pandas==1.3.1
    # via
    #   -r requirements-ds.in
    #   catboost
    #   google-cloud-bigquery
    #   pandas-gbq
    #   seaborn
    #   shap
    #   statsmodels
pandas_gbq==0.16.0
    # via -r requirements-ds.in
patsy==0.5.2
    # via statsmodels
pep517==0.12.0
    # via pip-tools
pillow==9.0.1
    # via matplotlib
pip-tools==6.2.0
    # via -r requirements-ds.in
platformdirs==2.5.1
    # via virtualenv
plotly==5.6.0
    # via catboost
pot==0.8.1.0
    # via -r requirements-ds.in
pre-commit==2.12.1
    # via -r requirements-ds.in
proto-plus==1.20.3
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
protobuf==3.19.4
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyarrow==5.0.0
    # via
    #   -r requirements-ds.in
    #   google-cloud-bigquery
    #   pandas-gbq
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydata-google-auth==1.4.0
    # via pandas-gbq
pyparsing==3.0.7
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2022.1
    # via pandas
pyyaml==6.0
    # via pre-commit
requests==2.27.1
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.8
    # via google-auth
scikit-learn==1.0.2
    # via
    #   lightgbm
    #   shap
scipy==1.7.3
    # via
    #   catboost
    #   lightgbm
    #   pot
    #   scikit-learn
    #   seaborn
    #   shap
    #   statsmodels
    #   xgboost
seaborn==0.11.2
    # via -r requirements-ds.in
shap==0.39.0
    # via -r requirements-ds.in
six==1.16.0
    # via
    #   catboost
    #   google-auth
    #   google-cloud-storage
    #   grpcio
    #   patsy
    #   plotly
    #   python-dateutil
    #   virtualenv
slicer==0.0.7
    # via shap
statsmodels==0.13.2
    # via -r requirements-ds.in
tenacity==8.0.1
    # via plotly
threadpoolctl==3.1.0
    # via scikit-learn
toml==0.10.2
    # via pre-commit
tomli==2.0.1
    # via pep517
tqdm==4.63.1
    # via shap
typing-extensions==4.3.0
    # via
    #   importlib-metadata
    #   kiwisolver
urllib3==1.26.9
    # via requests
virtualenv==20.14.0
    # via pre-commit
wheel==0.37.1
    # via
    #   lightgbm
    #   pip-tools
xgboost==1.4.2
    # via -r requirements-ds.in
zipp==3.8.1
    # via
    #   importlib-metadata
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
